docker/server/requirements.txt
qdrant_client==1.13.3
linkup_sdk==0.2.4
llama-index==0.12.27
llama-index-core==0.12.27
llama-index-llms-groq==0.3.1
llama-index-vector-stores-qdrant==0.6.0
llama-index-embeddings-huggingface==0.5.2
fastembed==0.6.0
mcp==1.6.0
python_dotenv==1.1.0